input: synaptics: s3320: Handle more framebuffer hints
The driver currently only tries to honor doze mode by watching
for one of the low power mode hints and assuming all other modes
are not doze.  Recent changes to Ambient display have taken
advantage of additional power states, which we wrongly assumed
should not be using the touchpanel.  Fix this by handling all of
the currently supported power states.

This is a port of one CyanogenOS commit to fix touchscreen with
ambient display. Actually it should be fixed by previous commit
as we allow to enable touchscreen in any case by removing bogus
workqueue which only handles UNBLANK case. But let's keep it
safe here.

Signed-off-by: Wang Han <416810799@qq.com>
Signed-off-by: Shreyansh Lodha <slodha96@gmail.com>