arm64: topology: Fix MPIDR-based detection
The original patch packs the cluster_id, in hope of
providing linear mapping when the affinity tree is balanced.

But this serialization generates wrong cluster ids.
Take a simple example of system with 2 Quad core clusters.
The mpidr_hash.shift_aff[1] will be 6 as you need minimum 2 bits
to represent aff[0]. So you will end up with second cluster with
id = 4 instead of 1.

The patch below applies on top of the original patch
(commit 05ee953)

1. Dropped mpidr_hash-related bits, in favor of simpler shift/or using MPIDR.
2. Also addressed Lorenzo's comment about redundant cluster_id==-1 check.

Refer to the discussion here:
[1] https://www.marc.info/?l=linux-arm-kernel&m=140123471318574&w=2

Change-Id: If52d405266f084585e0c97617e21fb5f2e2d1587
Acked-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Signed-off-by: Zi Shen Lim <zlim@broadcom.com>
Signed-off-by: Mark Brown <broonie@linaro.org>
Patch-mainline: linux-arm-kernel 5/27/2014 23:49
Signed-off-by: Srinivas Ramana <sramana@codeaurora.org>
Signed-off-by: Shreyansh Lodha <slodha96@gmail.com>